adding

+++ POC-II-Level/level2.sh
@@ -12,24 +12,13 @@ echo $total_size
 cd $line
 for line in $(find . -path '*/\.*' -prune -o -type f -exec du -sk {} + | sort -rn | head -10 | awk '{print $NF}')
         {
-        if [ -f "$line" ]; then
-            echo -e "$line is a file";
-            used_fsize=$(ls -lrt $line | awk '{print $5F}')
-            echo $total_size $used_fsize 
-            fpercent=$((100*$used_fsize/$to{tal_size ))
-                if [ $fpercent -ge 40 ]
-                {
-                    echo "$line is > 40%"
-                }
-                else {
-                    echo " " > /dev/null
-                fi
-                }
+         used_fsize=$(ls -lrt $line | awk '{print $5F}')
+        if [ -f "$line" && "$((100*$used_fsize/$total_size ))" -ge 40 ]; then
+            echo "$line is > 40%" 
         else
             echo " " > /dev/null
         fi
         } 
-     }
 <<comm
 level3_dftrack()
 {
